Whoops I'm a bad modeller
After a bit of work testing with modelling my own keyboard, I've decided I'm just going to cut that step, and model a basic frame of the keyboard instead. This will look bad, but it will look about the same as if I spent time modelling a higher quality one, this just saves me two weeks of work.